Backend/requirements.txt
# Core framework
fastapi==0.124.4
uvicorn==0.38.0
httptools>=0.7.0  # Better HTTP parsing for browser requests
SQLAlchemy==2.0.45
psycopg2-binary==2.9.11
pydantic==2.12.5
passlib==1.7.4
python-dotenv==1.2.1
PyJWT==2.10.1
redis==7.1.0
# For Argon2 password hashing
argon2-cffi
# Optional: for email validation (if needed)
email-validator
